Rename file and fix test
